able to get the bot subscribed to listen to channel point rewards.     currently have Plant, Water, and Attack.     Attack checks if a real twitch user is attacked and will mention
