+++ plans.md
@@ -6,5 +6,5 @@
     - Bypass timer for Water (1 time - has seperate cool down timer) - 50 bits?
     - Water someone elses plant - 50 bits?
     - Protect your plant from Super Attack for a set time (2 minutes?) - 100 bits
-    - Super attack (Attacks plant in front of you and switches spots - no cool down) - 200 bits
+    - Super attack (Attacks plant in front of you and switches spots IF the plant is thirsty - no cool down) - 200 bits
       - Tally how many attacks someone gives/receives
